# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flares",
    "description": "HTTP issue tracking with Pushover notifications",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/heartbeat": {
      "delete": {
        "tags": [],
        "operationId": "delete_heartbeat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Monitor deleted"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/heartbeats": {
      "get": {
        "tags": [],
        "operationId": "list_heartbeats",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Heartbeat"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "post": {
        "tags": [],
        "operationId": "register_heartbeat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Heartbeat"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/heartbeats/check-in": {
      "post": {
        "tags": [],
        "operationId": "check_in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseIssue"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Heartbeat"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CloseIssue": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
      "Heartbeat": {
        "allOf": [
          {
            "$ref": "#/components/schemas/HeartbeatInput"
          },
          {
            "type": "object",
            "required": [
              "last_seen",
              "due_at",
              "overdue"
            ],
            "properties": {
              "due_at": {
                "type": "integer",
                "format": "int64"
              },
              "last_seen": {
                "type": "integer",
                "format": "int64"
              },
              "overdue": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "HeartbeatInput": {
        "type": "object",
        "required": [
          "id",
          "title",
          "interval_seconds"
        ],
        "properties": {
          "grace_seconds": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "id": {
            "type": "string"
          },
          "interval_seconds": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "notify_on_recovery": {
            "type": "boolean"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "title": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "info",
          "warning",
          "critical"
        ]
      }
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}